# swagger.yml
...
openapi: 3.0.0
info:
  title: "ASCARIS REST API"
  description: "An API for ASCARIS, a digital semantic publishing application"
  version: "1.0.0"
servers:
  - url: "/api/v1/"
components:
  schemas:
...
    Document:
      type: "object"
      required:
        - id
      properties:
        docTitle:
          type: "string"
        content:
          type: "string"
        title:
          type: "array"
          items:
            $ref: "#/components/schemas/Title"
        responsibility:
          type: "array"
          items:
            $ref: "#/components/schemas/Responsibility"
        pubAuthority:
          type: "array"
          items:
            $ref: "#/components/schemas/PubAuthority"
        identifier:
          type: "array"
          items:
            $ref: "#/components/schemas/Identifier"
        source:
          type: "array"
          items:
            $ref: "#/components/schemas/Source"
        note:
          type: "array"
          items:
            $ref: "#/components/schemas/Note"
        category:
          type: "array"
          items:
            $ref: "#/components/schemas/Category"
        pubDate:
          type: "array"
          items:
            $ref: "#/components/schemas/PubDate"
        pubPlace:
          type: "array"
          items:
            $ref: "#/components/schemas/PubPlace"
        license:
          type: "array"
          items:
            $ref: "#/components/schemas/License"
        abstract:
          type: "array"
          items:
            $ref: "#/components/schemas/Abstract"
        creationDate:
          type: "array"
          items:
            $ref: "#/components/schemas/CreationDate"
  parameters:
    id:
      name: "id"
      description: "Identifier of the document to get"
      in: path
      required: True
      schema:
        type: "number"
    title_id:
      name: "title_id"
      description: "Identifier of the title"
      in: path
      required: True
      schema:
        type: "integer"
...
paths:
...
  /document/{id}:
    get:
      operationId: "document.read_one"
      tags:
        - Document
      summary: "Read one document"
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: "Success: read document"
    put:
      tags:
        - Document
      operationId: "document.update"
      summary: "Update a document"
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: "Success: updated document"
      requestBody:
        content:
          application/json:
            schema:
              x-body-name: "document"
              $ref: "#/components/schemas/Document"
    delete:
      tags:
        - Document
      operationId: "document.delete"
      summary: "Delete a document"
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "204":
          description: "Success: deleted document"
  /title:
    post:
      operationId: "title.create"
      tags:
        - Title
      summary: "Create a title associated with a document"
      requestBody:
        description: "Title to create"
        required: True
        content:
          application/json:
            schema:
              x-body-name: "title"
              type: "object"
              properties:
                document_id:
                  type: "integer"
                text:
                  type: "string"
                language:
                  type: "string"
      responses:
        "201":
          description: "Success: title created"
  /title/{title_id}:
    get:
      operationId: "title.read_one"
      tags:
        - Title
      summary: "Read a title"
      parameters:
        - $ref: "#/components/parameters/title_id"
      responses:
        "200":
          description: "Success: read title"
    put:
      tags:
        - Title
      operationId: "title.update"
      summary: "Update a title"
      parameters:
        - $ref: "#/components/parameters/title_id"
      responses:
        "200":
          description: "Success: title updated"
      requestBody:
        content:
          application/json:
            schema:
              x-body-name: "title"
              type: "object"
              properties:
                text:
                  type: "string"
                language:
                  type: "string"
    delete:
      tags:
        - Title
      operationId: "title.delete"
      summary: "Delete a title"
      parameters:
        - $ref: "#/components/parameters/title_id"
      responses:
        "204":
          description: "Success: title deleted"
  /responsibility:
    post:
      operationId: "responsibility.create"
      tags:
        - Responsibility
      summary: "Create a responsibility associated with a document"
      requestBody:
        description: "Responsibility to create"
        required: True
        content:
          application/json:
            schema:
              x-body-name: "responsibility"
              type: "object"
              properties:
                document_id:
                  type: "integer"
                surname:
                  type: "string"
                name:
                  type: "string"
                authority:
                  type: "string"
                role:
                  type: "string"
      responses:
        "201":
          description: "Success: responsibility created"
  /responsibility/{responsibility_id}:
    get:
      operationId: "responsibility.read_one"
      tags:
        - Responsibility
      summary: "Read a responsibility"
      parameters:
        - $ref: "#/components/parameters/responsibility_id"
      responses:
        "200":
          description: "Success: responsibility read"
    put:
      tags:
        - Responsibility
      operationId: "responsibility.update"
      summary: "Update a responsibility"
      parameters:
        - $ref: "#/components/parameters/responsibility_id"
      responses:
        "200":
          description: "Success: responsibility updated"
      requestBody:
        content:
          application/json:
            schema:
              x-body-name: "responsibility"
              type: "object"
              properties:
                surname:
                  type: "string"
                name:
                  type: "string"
                authority:
                  type: "string"
                role:
                  type: "string"
    delete:
      tags:
        - Responsibility
      operationId: "responsibility.delete"
      summary: "Delete a responsibility"
      parameters:
        - $ref: "#/components/parameters/responsibility_id"
      responses:
        "204":
          description: "Success: responsibility deleted"
  /pubAuthority:
    post:
      operationId: "pubAuthority.create"
      tags:
        - PubAuthority
      summary: "Create a pubAuthority associated with a document"
      requestBody:
        description: "pubAuthority to create"
        required: True
        content:
          application/json:
            schema:
              x-body-name: "pubAuthority"
              type: "object"
              properties:
                document_id:
                  type: "integer"
                name:
                  type: "string"
                authority:
                  type: "string"
                role:
                  type: "string"
      responses:
        "201":
          description: "Success: pubAuthority created"
  /pubAuthority/{pubAuthority_id}:
    get:
      operationId: "pubAuthority.read_one"
      tags:
        - PubAuthority
      summary: "Read a pubAuthority"
      parameters:
        - $ref: "#/components/parameters/pubAuthority_id"
      responses:
        "200":
          description: "Success: pubAuthority read"
    put:
      tags:
        - PubAuthority
      operationId: "pubAuthority.update"
      summary: "Update a pubAuthority"
      parameters:
        - $ref: "#/components/parameters/pubAuthority_id"
      responses:
        "200":
          description: "Success: pubAuthority updated"
      requestBody:
        content:
          application/json:
            schema:
              x-body-name: "pubAuthority"
              type: "object"
              properties:
                name:
                  type: "string"
                authority:
                  type: "string"
                role:
                  type: "string"
    delete:
      tags:
        - PubAuthority
      operationId: "pubAuthority.delete"
      summary: "Delete a pubAuthority"
      parameters:
        - $ref: "#/components/parameters/pubAuthority_id"
      responses:
        "204":
          description: "Success: pubAuthority deleted"
  /identifier:
    post:
      operationId: "identifier.create"
      tags:
        - Identifier
      summary: "Create a identifier associated with a document"
      requestBody:
        description: "identifier to create"
        required: True
        content:
          application/json:
            schema:
              x-body-name: "identifier"
              type: "object"
              properties:
                document_id:
                  type: "integer"
                text:
                  type: "string"
                language:
                  type: "string"
      responses:
        "201":
          description: "Success: identifier created"
  /identifier/{identifier_id}:
    get:
      operationId: "identifier.read_one"
      tags:
        - Identifier
      summary: "Read a identifier"
      parameters:
        - $ref: "#/components/parameters/identifier_id"
      responses:
        "200":
          description: "Success: read identifier"
    put:
      tags:
        - Identifier
      operationId: "identifier.update"
      summary: "Update a identifier"
      parameters:
        - $ref: "#/components/parameters/identifier_id"
      responses:
        "200":
          description: "Success: identifier updated"
      requestBody:
        content:
          application/json:
            schema:
              x-body-name: "identifier"
              type: "object"
              properties:
                text:
                  type: "string"
                language:
                  type: "string"
    delete:
      tags:
        - Identifier
      operationId: "identifier.delete"
      summary: "Delete a identifier"
      parameters:
        - $ref: "#/components/parameters/identifier_id"
      responses:
        "204":
          description: "Success: identifier deleted"
  /source:
    post:
      operationId: "source.create"
      tags:
        - Source
      summary: "Create a source associated with a document"
      requestBody:
        description: "source to create"
        required: True
        content:
          application/json:
            schema:
              x-body-name: "source"
              type: "object"
              properties:
                document_id:
                  type: "integer"
                text:
                  type: "string"
      responses:
        "201":
          description: "Success: source created"
  /source/{source_id}:
    get:
      operationId: "source.read_one"
      tags:
        - Source
      summary: "Read a source"
      parameters:
        - $ref: "#/components/parameters/source_id"
      responses:
        "200":
          description: "Success: read source"
    put:
      tags:
        - Source
      operationId: "source.update"
      summary: "Update a source"
      parameters:
        - $ref: "#/components/parameters/source_id"
      responses:
        "200":
          description: "Success: source updated"
      requestBody:
        content:
          application/json:
            schema:
              x-body-name: "source"
              type: "object"
              properties:
                text:
                  type: "string"
    delete:
      tags:
        - Source
      operationId: "source.delete"
      summary: "Delete a source"
      parameters:
        - $ref: "#/components/parameters/source_id"
      responses:
        "204":
          description: "Success: source deleted"
  /note:
    post:
      operationId: "note.create"
      tags:
        - Note
      summary: "Create a note associated with a document"
      requestBody:
        description: "note to create"
        required: True
        content:
          application/json:
            schema:
              x-body-name: "note"
              type: "object"
              properties:
                document_id:
                  type: "integer"
                text:
                  type: "string"
      responses:
        "201":
          description: "Success: note created"
  /note/{note_id}:
    get:
      operationId: "note.read_one"
      tags:
        - Note
      summary: "Read a note"
      parameters:
        - $ref: "#/components/parameters/note_id"
      responses:
        "200":
          description: "Success: read note"
    put:
      tags:
        - Note
      operationId: "note.update"
      summary: "Update a note"
      parameters:
        - $ref: "#/components/parameters/note_id"
      responses:
        "200":
          description: "Success: note updated"
      requestBody:
        content:
          application/json:
            schema:
              x-body-name: "note"
              type: "object"
              properties:
                text:
                  type: "string"
    delete:
      tags:
        - Note
      operationId: "note.delete"
      summary: "Delete a note"
      parameters:
        - $ref: "#/components/parameters/note_id"
      responses:
        "204":
          description: "Success: note deleted"
  /category:
    post:
      operationId: "category.create"
      tags:
        - Category
      summary: "Create a category associated with a document"
      requestBody:
        description: "category to create"
        required: True
        content:
          application/json:
            schema:
              x-body-name: "category"
              type: "object"
              properties:
                document_id:
                  type: "integer"
                type:
                  type: "string"
      responses:
        "201":
          description: "Success: category created"
  /category/{category_id}:
    get:
      operationId: "category.read_one"
      tags:
        - Category
      summary: "Read a category"
      parameters:
        - $ref: "#/components/parameters/category_id"
      responses:
        "200":
          description: "Success: read category"
    put:
      tags:
        - Category
      operationId: "category.update"
      summary: "Update a category"
      parameters:
        - $ref: "#/components/parameters/category_id"
      responses:
        "200":
          description: "Success: category updated"
      requestBody:
        content:
          application/json:
            schema:
              x-body-name: "category"
              type: "object"
              properties:
                type:
                  type: "string"
    delete:
      tags:
        - Category
      operationId: "category.delete"
      summary: "Delete a category"
      parameters:
        - $ref: "#/components/parameters/category_id"
      responses:
        "204":
          description: "Success: category deleted"
  /pubDate:
    post:
      operationId: "pubDate.create"
      tags:
        - PubDate
      summary: "Create a pubDate associated with a document"
      requestBody:
        description: "pubDate to create"
        required: True
        content:
          application/json:
            schema:
              x-body-name: "pubDate"
              type: "object"
              properties:
                document_id:
                  type: "integer"
                date:
                  type: "string"
                  format: "date-time"
      responses:
        "201":
          description: "Success: pubDate created"
  /pubDate/{pubDate_id}:
    get:
      operationId: "pubDate.read_one"
      tags:
        - PubDate
      summary: "Read a pubDate"
      parameters:
        - $ref: "#/components/parameters/pubDate_id"
      responses:
        "200":
          description: "Success: read pubDate"
    put:
      tags:
        - PubDate
      operationId: "pubDate.update"
      summary: "Update a pubDate"
      parameters:
        - $ref: "#/components/parameters/pubDate_id"
      responses:
        "200":
          description: "Success: pubDate updated"
      requestBody:
        content:
          application/json:
            schema:
              x-body-name: "pubDate"
              type: "object"
              properties:
                date:
                  type: "string"
                  format: "date-time"
    delete:
      tags:
        - PubDate
      operationId: "pubDate.delete"
      summary: "Delete a pubDate"
      parameters:
        - $ref: "#/components/parameters/pubDate_id"
      responses:
        "204":
          description: "Success: pubDate deleted"
  /pubPlace:
    post:
      operationId: "pubPlace.create"
      tags:
        - PubPlace
      summary: "Create a pubPlace associated with a document"
      requestBody:
        description: "pubPlace to create"
        required: True
        content:
          application/json:
            schema:
              x-body-name: "pubPlace"
              type: "object"
              properties:
                document_id:
                  type: "integer"
                name:
                  type: "string"
                authority: 
                  type: "string"
      responses:
        "201":
          description: "Success: pubPlace created"
  /pubPlace/{pubPlace_id}:
    get:
      operationId: "pubPlace.read_one"
      tags:
        - PubPlace
      summary: "Read a pubPlace"
      parameters:
        - $ref: "#/components/parameters/pubPlace_id"
      responses:
        "200":
          description: "Success: read pubPlace"
    put:
      tags:
        - PubPlace
      operationId: "pubPlace.update"
      summary: "Update a pubPlace"
      parameters:
        - $ref: "#/components/parameters/pubPlace_id"
      responses:
        "200":
          description: "Success: pubPlace updated"
      requestBody:
        content:
          application/json:
            schema:
              x-body-name: "pubPlace"
              type: "object"
              properties:
                name:
                  type: "string"
                authority: 
                  type: "string"
    delete:
      tags:
        - PubPlace
      operationId: "pubPlace.delete"
      summary: "Delete a pubPlace"
      parameters:
        - $ref: "#/components/parameters/pubPlace_id"
      responses:
        "204":
          description: "Success: pubPlace deleted"
  /license:
    post:
      operationId: "license.create"
      tags:
        - License
      summary: "Create a license associated with a document"
      requestBody:
        description: "license to create"
        required: True
        content:
          application/json:
            schema:
              x-body-name: "license"
              type: "object"
              properties:
                document_id:
                  type: "integer"
                text:
                  type: "string"
                link: 
                  type: "string"
      responses:
        "201":
          description: "Success: license created"
...
  /abstract:
    post:
      operationId: "abstract.create"
      tags:
        - Abstract
      summary: "Create a abstract associated with a document"
      requestBody:
        description: "abstract to create"
        required: True
        content:
          application/json:
            schema:
              x-body-name: "abstract"
              type: "object"
              properties:
                document_id:
                  type: "integer"
                text:
                  type: "string"
      responses:
        "201":
          description: "Success: abstract created"
...